corrected

+++ RangeQueries/Static_Range_Sum_Queries.cpp
@@ -23,7 +23,7 @@ const int mod = 1e9 + 7;
 const ll inf = 1e18;
 
 const int N = 2e5 + 2;
-int arr[N];
+ll arr[N];
 ll prefix[N];
 
 void solve()
